imagen en la categoria, varias imagenes en el articulo

+++ src/category/category.service.ts
@@ -1,28 +1,66 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { CreateCategoryDto } from './dto/create-category.dto';
 import { UpdateCategoryDto } from './dto/update-category.dto';
 import { Category } from './entities/category.entity';
+import { CloudinaryService } from 'src/cloudinay/cloudinay.service';
 
 @Injectable()
 export class CategoryService {
   constructor(
     @InjectRepository(Category)
     private readonly categoryRepository: Repository<Category>,
-  ) {}
+    private cloudinaryService: CloudinaryService,
+  ) { }
 
-  async create(
-    createCategoryDto: CreateCategoryDto,
-  ): Promise<{ message: string; category: Category }> {
-    const category = this.categoryRepository.create(createCategoryDto);
-    const savedCategory = await this.categoryRepository.save(category);
+  // async create(
+  //   createCategoryDto: CreateCategoryDto,
+  // ): Promise<{ message: string; category: Category }> {
+  //   const category = this.categoryRepository.create(createCategoryDto);
+  //   const savedCategory = await this.categoryRepository.save(category);
+
+  //   return {
+  //     message: 'Categoría creada con éxito',
+  //     category: savedCategory,
+  //   };
+  // }
+
+  async create(categoria: CreateCategoryDto) {
+
+    const categoriaFound = await this.categoryRepository.findOne({
+      where: { nombre: categoria.nombre },
+    });
+
+    if (categoriaFound) {
+      throw new HttpException('La categoría ya existe.', HttpStatus.CONFLICT);
+    }
+
+    const uploadResult = await this.cloudinaryService.uploadFile(
+      categoria.imagen,
+    );
+
+    if (!uploadResult) {
+      throw new HttpException(
+        'Error al subir la imagen.',
+        HttpStatus.INTERNAL_SERVER_ERROR,
+      );
+    }
+
+    const imageUrl = uploadResult.secure_url;
+
+    const newCategoria = this.categoryRepository.create({
+      ...categoria,
+      imagen: imageUrl,
+    });
+
+    await this.categoryRepository.save(newCategoria);
 
     return {
       message: 'Categoría creada con éxito',
-      category: savedCategory,
     };
   }
+
 
   async findAll(): Promise<Category[]> {
     return this.categoryRepository.find();
@@ -57,12 +95,58 @@ export class CategoryService {
     };
   }
 
-  async remove(id: number): Promise<{ message: string }> {
-    const { category } = await this.findOne(id);
-    await this.categoryRepository.remove(category);
+  private extractPublicId(imageUrl: string): string | null {
+    const regex = /\/([^\/]+)\.\w+$/;
+    const match = imageUrl.match(regex);
+    return match ? match[1] : null;
+  }
+
+  async updateImagen(
+    id: any,
+    imagen: Express.Multer.File,
+  ): Promise<{ message: string; category: Category }> {
+    const category = await this.categoryRepository.findOne({ where: { id } });
+
+    if (!category) {
+      throw new NotFoundException(`Categoria no encontrado`);
+    }
+
+    const currentImageUrl = category.imagen;
+    if (currentImageUrl) {
+      const publicId = this.extractPublicId(currentImageUrl);
+      if (publicId) {
+        await this.cloudinaryService.deleteFile(publicId);
+      }
+    }
+
+    const uploadResult = await this.cloudinaryService.uploadFile(imagen);
+
+    category.imagen = uploadResult.secure_url;
+
+    const updatedCategory = await this.categoryRepository.save(category);
 
     return {
-      message: `Categoría eliminada con éxito`,
+      message: 'Imagen actualizada con éxito',
+      category: updatedCategory,
     };
+  }
+
+  async remove(id: any): Promise<{ message: string }> {
+    const category = await this.categoryRepository.findOne({ where: { id } });
+
+    if (!category) {
+      throw new NotFoundException(`Categoría no encontrado`);
+    }
+
+    const imageUrl = category.imagen;
+    const publicId = this.extractPublicId(imageUrl);
+
+    if (publicId) {
+      await this.cloudinaryService.deleteFile(publicId);
+    }
+
+    await this.categoryRepository.remove(category);
+
+    return { message: 'Artículo eliminado con éxito' };
   }
 }
